Implement container resizing
Resizing of containers now works as well as it is going to. I need to
implement a method that will calculate the height of the Container if a
width and height are not provided - this simply requires generating a
width and height, and then checking to see if any components have a
right / bottom that extends outside the current width / height, and
simply extending the Container if that is the case.  The redraw() method
in the Exchange class is called once the program is run such that an
initial graphical state is guaranteed.

The Percents class needed to cast integers to doubles when calculating
percentages so that has been fixed.

The TestModule class is simply that, a test Module that I am using to
test various features. Next I am going to make sure focus change works
right.

I think I need to make an onFocus() and maybe onUnfocus() method for the
Module class so that that works out nicely.

+++ vague/util/Percents.java
@@ -32,14 +32,15 @@ public class Percents {
      * Creates a Percents object based on the values for a percentage.
      * 
      * This constructor uses integer values.
-     * @param partialWidth The width of the part.
-     * @param totalWidth The width of the whole.
-     * @param partialHeight The height of the part.
-     * @param totalHeight The height of the whole.
+     * @param partX The x of the part.
+     * @param totalX The x of the whole.
+     * @param partY The y of the part.
+     * @param totalY The y of the whole.
      */
     public Percents(int partX, int totalX, int partY, int totalY) {
-        x = partX / totalX;
-        y = partY / totalY;
+        //Cast to double so percent != 0 
+        x = (double)partX / totalX;
+        y = (double)partY / totalY;
         initialized = true;
     }
     
@@ -47,10 +48,10 @@ public class Percents {
      * Creates a Percents object based on the values for a percentage.
      * 
      * This constructor uses floating point values.
-     * @param partialWidth The width of the part.
-     * @param totalWidth The width of the whole.
-     * @param partialHeight The height of the part.
-     * @param totalHeight The height of the whole.
+     * @param partX The x of the part.
+     * @param totalX The x of the whole.
+     * @param partY The y of the part.
+     * @param totalY The y of the whole.
      */
     public Percents(double partX, double totalX, double partY, double totalY) {
         x = partX / totalX;
@@ -69,8 +70,9 @@ public class Percents {
     }
     
     public void update(int partX, int totalX, int partY, int totalY) {
-        x = partX / totalX;
-        y = partY / totalY;
+        //Cast to double so percent != 0
+        x = (double)partX / totalX;
+        y = (double)partY / totalY;
         initialized = true;
     }
